+++ compiler/src/main/java/info/teksol/mc/mindcode/compiler/optimization/DataFlowOptimizer.java
@@ -304,6 +304,7 @@ class DataFlowOptimizer extends BaseOptimizer {
         }
 
         VariableStates variableStates = dataFlowVariableStates.createVariableStates();
+        variableStates.markInitialized(LogicVariable.remoteWaitAddr());
         if (context.isFunction()) {
             MindcodeFunction function = context.function();
             assert function != null;