+++ schemacode/src/main/java/info/teksol/schemacode/mindustry/Implementation.java
@@ -110,6 +110,7 @@ public enum Implementation {
     REPAIRTOWER,
     REPAIRTURRET,
     ROUTER,
+    RUNEOVERLAY,
     SEABUSH,
     SEAWEED,
     SEPARATOR,